修改 sqlite 类型

+++ ToolGood.ReadyGo3/Gadget/TableManager/Providers/SQLiteDatabaseProvider.cs
@@ -150,9 +150,9 @@ namespace ToolGood.ReadyGo3.Gadget.TableManager.Providers
             if (type == typeof(Int16)) return CreateField(ti, ci, "INTEGER", ci.FieldLength, isRequired);
             if (type == typeof(Int32)) return CreateField(ti, ci, "INTEGER", ci.FieldLength, isRequired);
             if (type == typeof(Int64)) return CreateField(ti, ci, "INTEGER", ci.FieldLength, isRequired);
-            if (type == typeof(Single)) return CreateField(ti, ci, "double", ci.FieldLength, isRequired);
-            if (type == typeof(double)) return CreateField(ti, ci, "double", ci.FieldLength, isRequired);
-            if (type == typeof(decimal)) return CreateField(ti, ci, "double", ci.FieldLength, isRequired);
+            if (type == typeof(Single)) return CreateField(ti, ci, "REAL", ci.FieldLength, isRequired);
+            if (type == typeof(double)) return CreateField(ti, ci, "REAL", ci.FieldLength, isRequired);
+            if (type == typeof(decimal)) return CreateField(ti, ci, "REAL", ci.FieldLength, isRequired);
             if (type == typeof(DateTime)) return CreateField(ti, ci, "dateTime", ci.FieldLength, isRequired);
             if (type == typeof(TimeSpan)) return CreateField(ti, ci, "dateTime", ci.FieldLength, isRequired);
             if (type == typeof(DateTimeOffset)) return CreateField(ti, ci, "dateTime", ci.FieldLength, isRequired);
